add replay logic collect the command, going to excuted the command to replay next

+++ src/aCommand.java
@@ -1,5 +1,5 @@
 public class aCommand extends Command{
-    public aCommand(Character character, long time){
+    public aCommand(Character character, double time){
         super(character, time);
     }
